[US_PA][Workflows] Include program names in sex offender criterion (Recidiviz/recidiviz-data#38102)
## Description of the change

This expands the `not_on_sex_offense_protocol` criterion to look at sex
offender related program names, instead of just program codes.

Validation: Generated a sheet
[here](https://docs.google.com/spreadsheets/d/1EcH6fFmbENbTUAkP5RfHYGkM2mOo8wmdQPZ3arfesY0/edit?gid=77747564#gid=77747564)
with all of the new program codes/names that will be considered as sex
offense related to sanity check. Also looked at changes in eligibility -
as expected, this is a minor change which will result in 13 less people
being eligible for special circ (out of ~4000) and 28 less people being
eligible for admin (out of ~2700).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e5b169b26666723cba0910d8f024763d1d8786d0