[Utils] Remove specific entity module imports from entity_utils (Part 1) (Recidiviz/recidiviz-data#38288)
## Description of the change

Introduces the `EntitiesModuleContext` interface, which is essentially a
wrapper around a `ModuleType` module and the associated class hierarchy
for the classes in that module. In a *separate* file, defines context
objects for each specific entities module. This sets us up for a
refactor in which we'll eventually be able to eliminate direct imports
to `state_entities` etc to `entity_utils.py` by instead passing in the
module context.

Also moves `EntityFieldIndex` into its own file so we can expose it via
the `EntitiesModuleContext` without creating a circular import with
`entity_utils.py`.

The eventual goal of this change is to be able to reference code in
`entities.py` in our entities modules (e.g. reference serialization
utils).

Subsequent PRs will address the TODOs I added in `entity_utils.py`.
There are just a lot of callsites to update and it would be too much for
one PR.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#38281

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 62a8019a9827f40c65bb60872b1e3f9829b5a667