update previous-if-transitional metrics to regular when it makes sense to (Recidiviz/recidiviz-data#38111)
## Description of the change

We only be reading from the previous-if-transitional table for outcomes
metrics, and should default to the regular officer agg metrics table for
everything else.

Looking for reviews from:
* @Gina-Valderrama for vitals, outliers_views, and insights perf review
changes
* @mayukas for misc agg metrics, agg metrics utils, and looker changes

@ageiduschek happy to have a review wherever you feel like you can add
something useful if you feel like reviewing since you've been involved
in this project
@samanthahuff fyi, this includes the change to use regular officer
metrics to determine exclusions

Also looking for a review of what's **missing** from this PR. After this
PR, the places that still will be using previous-if-transitional metrics
that aren't present in this PR are:
* aggregated metrics infra (assignment sessions, standard collection,
tests) (@mayukas or @ageiduschek )
* supervision_client_events for insights: this is just using assignment
sessions to determine the correct officer, and since these events, in
theory, contribute to the outcomes metrics it seemed like it made sense
to associate them with that officer (@Gina-Valderrama, @mayukas, or
@samanthahuff )
* insights perf review sheets, the part that pulls outcomes data (this
doesn't super matter since the project is done) (@Gina-Valderrama )
* custom_insights_metrics_lookml_generator: since this is for insights
impact and some of the metrics involved are outcomes ones it felt like
maybe it should still use the previous-if-transitional table, but this
i'm less confident about (@mayukas )

I don't have this in a sandbox yet because I want the views to update
from today's staging deploy first so I can more easily load what's
changed and compare. Will update once that's ready.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

https://github.com/Recidiviz/recidiviz-data/issues/37893

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 4c0a1ef267062b2eff130b0ecef4714de0899ceb