[Aggregated Metrics] Address a few misc remaining TODOs (Recidiviz/recidiviz-data#37790)
## Description of the change

Addresses a handful of remaining easy TODOs for Recidiviz/recidiviz-data#29291:

1) Stops warning in `load_views_to_sandbox` for all views in the
`aggregated_metrics` dataset. The only remaining legacy views in this
dataset are the `MiscAggregatedMetrics` views and those are not really
expensive to load. Also removes the custom prompting for dataflow
metrics views - none of these views are even in the top 10 most
expensive views at this point. There is a world someday where we'd be
able to estimate how expensive a sandbox will be to load based on past
runs and warn based on that number. For now, it doesn't seem worth the
extra complexity.

2) Removes an already addressed exemption in
`test_no_expensive_union_all_view_queries` and fixes a bug that was
preventing us from failing when we didn't do that cleanup before.

3) Updates the `get_person_events()` notebook helper to use an
observation-specific view in its query.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Part of Recidiviz/recidiviz-data#29291

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d532c15996b156d92533c11d48a2a1305497ab98