Migrating next batch of TES criteria to sentencing V2 (Recidiviz/recidiviz-data#37985)
## Description of the change

This PR migrates a set of 11 TES criteria queries to the sentencing_v2
infra. The 11 criteria migrated are those flagged as "In Progress" in
[this](https://www.notion.so/recidiviz/d1a4da82396d45a7907243e718d8187c?v=960ddb57385944b78c0d0e4362200ee3)
doc.

My steps to validate the output following this changes is in
[this](https://colab.research.google.com/drive/1FJF30jN4K-pNfN9GFG4Gt_9nx-PIL_Er?usp=sharing)
notebook. The following are the steps that I took:
1. Determine the set of tasks that have dependencies on these 11
criteria
2. First validate at the criteria level and look for changes in the
current population that meet / do not meet each criteria
3. Then validate at the task level and look for changes in the current
population eligible / not eligible to confirm that any changes at the
criteria level do not have downstream impacts.

Validation summary: the 11 criteria are referenced by 7 tasks, and there
is essentially no impact at the task level (1 person went from eligible
--> not eligible and 1 person went from not eligible to eligible).

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Advances Recidiviz/recidiviz-data#33748

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 61fcd3d6a3a826f4906890885623ff1f752a8a16