Update cohort plotting function to return figure and axes objects (Recidiviz/recidiviz-data#38344)
## Description of the change

Makes a small update to the `plot_event_curves` function in
`cohort_methods.py` to return the matplotlib figure and axes objects so
that the plot can be further edited after being generated.

The function currently shows the plot without returning any object - any
notebook that previously uses the function should still run as it did
previously.

[This](https://colab.research.google.com/drive/1nv2B9J605DlGyokNlJcgcfJZUDZRXGqP?usp=sharing)
colab notebook demos the use of the function with this change.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: d4e9d414743e5b7ca862d71d17a7f4a4ae3d6eaa