[Sentencing] Make county `null` instead of `UNKNOWN` when it is unknown or out of state (Recidiviz/recidiviz-data#38148)
## Description of the change
Right now in Idaho we label a county as "UNKNOWN" when it is unknown,
but we should just make it null since we support that on the database
side.

Similarly, we set a county of "OUT OF STATE" in North Dakota when it is
out of state, but just set to `null` for now until we have a better user
flow for out of state cases.

Tested this out with the load_datasets_to_sandbox +
export_metrics_from_dataset_to_gcs tools and everything looks correct!

FYI @mxosman Makes
https://github.com/Recidiviz/recidiviz-dashboards/pull/7300#issuecomment-2643273633
and https://github.com/Recidiviz/recidiviz-dashboards/pull/7318
unnecessary!

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 04cbd531f7750983293aef3bdba828f3d9bf52ab