[Testing] Update Ingest Structure Tests To Prep for Janet Implementation (Recidiviz/recidiviz-data#38325)
## Description of the change

This PR updates our existing ingest structure tests to account for all
of the changes that are about to happen. The full implementation plan
for "phase 1", the refactoring of our ingest tests before automating
fixture management is
[here](https://www.notion.so/recidiviz/Ingest-Test-Refinement-and-Fixture-Management-Janet-436961728cc44f6b968ba6fa387d67df?pvs=4Recidiviz/recidiviz-data#1967889f4d1980ff8ae8c3344bb6374d).

The issues of that plan are noted as TODOs in this PR and the new test.

The next PR will be to address Recidiviz/recidiviz-data#38323. I will then
rebase/update/incorporate feedback on the currently open PRs and merge
those in.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| **Type: Feature** | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38236

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 78d86876649137f84e034f2e803e5611f67872c3