[Justice Counts] Remove CSG From Reporting Agency Options (Recidiviz/recidiviz-data#38338)
This PR removes the CSG Justice Center as an option from the reporting
agency dropdown. Based on the latest requirements, CSG should not be
included as a valid selection.

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7c0b08290d1a5cc9932eed14acc48c919210a5a4