Add support for specifying required columns in `BigQueryView` (Recidiviz/recidiviz-data#37931)
## Description of the change
Add support for specifying the schema for the materialized view in a
`BigQueryView` object. Vertex Agent Builder requires that columns in a
table feeding into a datastore have REQUIRED columns, so that's why I'm
adding this.

This also is a good starting point for the work necessary for
https://github.com/Recidiviz/recidiviz-data/issues/18306!

Tested it with the view I want to update and it works!
<img width="384" alt="Screenshot 2025-02-11 at 4 13 49 PM"
src="https://github.com/user-attachments/assets/59d920da-9253-4262-a878-f4252c4d6099"
/>

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [x] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [x] This pull request has a descriptive title and information useful
to a reviewer
- [x] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: 7b8f0ad98915b6423e2f39f82525e5225a98cfff