[Testing] Have raw data fixtures have metadata raw data columns (Recidiviz/recidiviz-data#38360)
## Description of the change

Part of the challenge of using raw data fixtures for integration tests
was the schema difference between "ALL" and "LATEST" filter type
fixtures. In reality we use the full schema from "ALL" files, which has
metadata columns (the schema for `us_xx_raw_data`).

This PR updates all fixtures to have these columns, updates the loader
to check file ID and update datetime are compatible, and updates our
fixture generation script.

This PR does NOT unify the latest and all versions of fixtures. That
will be in a future PR and is noted as a TODO.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#38175

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: a427884319f899734056140dd716aa3dcd655953