[US_PA][Workflows] Update ineligible offenses for admin supervision (Recidiviz/recidiviz-data#37758)
## Description of the change

This PR updates logic around PA offense codes. Sorry it's kind of a lot
for one PR so going to break it down as best I can:

1. In the last round of TT calls, we found that we were using an
outdated version of form 402, which lays out the ineligible offenses for
admin supervision (see old version
[here](https://docs.google.com/document/d/1oyK0ykjfQPFe5ARZGbou6jEzQDFPt-9Y/edit)
and new version
[here](https://drive.google.com/file/d/1WLLvsGuKPUB4-JycX0KPLv6jqqvnESCi/view?usp=sharing)).
This PR changes the list of offenses in
`not_serving_ineligible_offense_for_admin_supervision` accordingly
2. We also got some additional feedback from TTs on which offenses
should be included as ineligible or
[violent](https://drive.google.com/file/d/1xqkTSMDbj7jU-htkp-1c-qtrYri-dhT8/view).
This PR expands `not_serving_ineligible_offense_for_admin_supervision`
and `offense_is_violent` to address this feedback as best I can
3. Finally, we ingested a new table of criminal history resulting in new
statutes/descriptions in `sentences_preprocessed_materialized`. The new
criminal history table had a lot more charges from other states, with
statute codes in different formats. This PR introduces the
`statute_code_is_like` function to check that statute codes match the PA
format and implements this in a few places where PA charges are used.
4. Previously, I required both the statute and description to match what
was expected if both were non-null. Now because there's a lot more
out-of-state offenses with non-null statutes, I'm relaxing this a bit to
count an offense if _either_ the statute or description matches.

Validation performed:

1. Ran a sandbox (prefix soph2) to look at eligibility - this PR leads
to a ~10% reduction in admin supervision eligibility (2763 -> 2488) and
a 0% reduction in special circ eligibility (4028 -> 4025).
2. Created a sheet listing all the offenses that are newly considered
admin ineligible / violent
[here](https://docs.google.com/spreadsheets/d/1t27iHuhnLOphUT4meGWJvOvJtQI0-SuhmLug_-IW5qU/edit?gid=1934876936#gid=1934876936)
to sanity check.

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes Recidiviz/recidiviz-data#37212, Closes Recidiviz/recidiviz-data#37213

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [x] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: e74535fd0cbaafe62a1ef0401033821d9b9b777d