[US_UT] Fix bug in `sentence_status` view (Recidiviz/recidiviz-data#38381)
## Description of the change

There are a very small number of cases where a `COMPLETED` sentence
status is followed by a `SERVING` status in Utah. This happens when a
case's termination date is populated and later unpopulated in raw data.
We do not know yet why that might be the case, so for the time being, we
are filtering these cases out so that ingest can run.

The way I was filtering them before was buggy - this fixes the logic so
that once a sentence has a `COMPLETED` status, that is guaranteed to be
the final status. This is not elegant but will be cleaned up with the
sentence refactor I'm actively working on - just trying to get the
ingest pipeline unblocked in the meantime!

[Sandbox
pipeline](https://console.cloud.google.com/dataflow/jobs/us-west1/2025-02-14_07_39_30-545544039354491959;graphView=1?project=recidiviz-staging&inv=1&invt=Abpjzg&pageState=(%22dfTime%22:(%22l%22:%22dfJobMaxTime%22)))

## Type of change

> All pull requests must have at least one of the following labels
applied (otherwise the PR will fail):

| Label | Description |
|-----------------------------
|-----------------------------------------------------------------------------------------------------------
|
| Type: Bug | non-breaking change that fixes an issue |
| Type: Feature | non-breaking change that adds functionality |
| Type: Breaking Change | fix or feature that would cause existing
functionality to not work as expected |
| Type: Non-breaking refactor | change addresses some tech debt item or
prepares for a later change, but does not change functionality |
| Type: Configuration Change | adjusts configuration to achieve some end
related to functionality, development, performance, or security |
| Type: Dependency Upgrade | upgrades a project dependency - these
changes are not included in release notes |

## Related issues

Closes #XXXX

## Checklists

### Development

**This box MUST be checked by the submitter prior to merging**:
- [ ] **Double- and triple-checked that there is no Personally
Identifiable Information (PII) being mistakenly added in this pull
request**

These boxes should be checked by the submitter prior to merging:
- [ ] Tests have been written to cover the code changed/added as part of
this pull request

### Code review

These boxes should be checked by reviewers prior to merging:

- [ ] This pull request has a descriptive title and information useful
to a reviewer
- [ ] Potential security implications or infrastructural changes have
been considered, if relevant

GitOrigin-RevId: f2c3e932e04700424f5c48bd36127b2a0225a124